removing bf16 completely even from quntization

+++ parallel_finetune.py
@@ -61,8 +61,10 @@ class Trainer:
             load_in_4bit=True,
             bnb_4bit_use_double_quant=True,
             bnb_4bit_quant_type="nf4",
-            bnb_4bit_compute_dtype=torch.bfloat16,
-            bnb_4bit_quant_storage=torch.bfloat16,
+            # bnb_4bit_compute_dtype=torch.bfloat16,
+            # bnb_4bit_quant_storage=torch.bfloat16,
+            bnb_4bit_compute_dtype=torch.float32,
+            bnb_4bit_quant_storage=torch.float32,
         )
     
     def _get_peft_config(self):
